[ISSUE-114] Refactor brain to remove CaslezoneSet references
