+++ brain/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "myopic-brain"
-version = "1.7.1"
+version = "1.8.0"
 authors = ["Thomas Ball <tomoliverball@gmail.com>"]
 edition = "2018"
 license = "MIT"
@@ -14,7 +14,7 @@ include = [
 ]
 
 [dependencies]
-myopic-board = "2.1.2"
+myopic-board = { path = "../board" }
 lazy_static = "1.4.0"
 itertools = "0.9.0"
 regex = "1"
