+++ brain/src/eval/clearcut.rs
